refactor : schedule / register 등록 로직 훅으로 분리
